move TagInput into dir, use button on desktop
